+++ src/common/validateRules.jsx
@@ -1,5 +1,6 @@
 import { getStrBytes, filterSymbol } from './utils';
 
+
 const checkMaxBytes = (rule, value, callback) => {
     if(getStrBytes(value) > rule.maxBytes){
         callback(rule.message);
@@ -44,18 +45,13 @@ const validateRules = {
     email: {type: 'email', message: '请输入有效的email地址'},
     date: {type: 'date', message: '请输入有效的日期'},
     number: {pattern: /^(?:-?\d+|-?\d{1,3}(?:,\d{3})+)?(?:\.\d+)?$/, message: '请输入有效数字'},
-    maxBytes: (num) => {
-        return { maxBytes: num, message: `该项字节长度不能超过${num}字节`, validator: checkMaxBytes};
-    },
-    minBytes: (num) => {
-        return { minBytes: num, message: `该项字节长度至少${num}字节`, validator: checkMinBytes};
-    },
-    maxBytesWithFilter: (num) => {
-        return { maxBytes: num, message: `该项字节长度不能超过${num}字节`, validator: checkMaxBytesWithFilter};
-    },
-    minBytesWithFilter: (num) => {
-        return { minBytes: num, message: `该项字节长度至少${num}字节`, validator: checkMinBytesWithFilter};
-    }
+    color: {pattern: /^#[0-9a-fA-F]{6}$/, message: '请输入有效的颜色值，格式：#FFFFFF'},
+    maxChars: (max) => ({type: 'string', max, message: `该项字符长度不能超过${max}字符`}),
+    minChars: (min) => ({type: 'string', min, message: `该项字符长度不能少于${min}字符`}),
+    maxBytes: (number) => ({maxBytes: number, message: `该项字节长度不能超过${number}字节`, validator: checkMaxBytes}),
+    minBytes: (number) => ({minBytes: number, message: `该项字节长度至少${number}字节`, validator: checkMinBytes}),
+    maxBytesWithFilter: (number) => ({maxBytes: number, message: `该项字节长度不能超过${number}字节`, validator: checkMaxBytesWithFilter}),
+    minBytesWithFilter: (number) => ({minBytes: number, message: `该项字节长度至少${number}字节`, validator: checkMinBytesWithFilter}),
 };
 
 export default validateRules;